feat: new routing (#76)
@race-of-sloths
