+++ src/features/sponsor-dashboard/components/Submissions/SubmissionHeader.tsx
@@ -67,7 +67,7 @@ export const SubmissionHeader = ({
 
   const deadline = formatDeadline(bounty?.deadline, bounty?.type);
 
-  const listingPath = `listing/${bounty?.slug}`;
+  const listingPath = `${bounty?.sponsor?.slug}/${bounty?.sequentialId}`;
   const { hasCopied, onCopy } = useClipboard(`${getURL()}${listingPath}`);
 
   const bountyStatus = getListingStatus(bounty);
@@ -75,7 +75,7 @@ export const SubmissionHeader = ({
   const listingLink =
     bounty?.type === 'grant'
       ? `${getURL()}grants/${bounty.slug}/`
-      : `${getURL()}listing/${bounty?.slug}/`;
+      : `${getURL()}${bounty?.sponsor?.slug}/${bounty?.sequentialId}/`;
 
   const socialListingLink = (medium?: 'twitter' | 'telegram') =>
     `${listingLink}${medium ? `?utm_source=${PROJECT_NAME}&utm_medium=${medium}&utm_campaign=sharelisting/` : ``}`;
